# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Quest API
  description: |
    REST API for Quest - a Jira-like project and task management system.

    ## Authentication
    All endpoints require authentication via session cookie (NextAuth.js) or Bearer token.

    ## Base URL
    - Production: `https://quest-manas-pradhans-projects.vercel.app`
    - Development: `http://localhost:3000`

    ## Notes for Mobile Developers
    This API specification documents the full capabilities of Quest. All endpoints are
    implemented REST API routes that wrap Next.js Server Actions, making them consumable
    by iOS, Android, and other platforms via standard HTTP requests.
  version: 1.0.0
  contact:
    name: Quest Development Team
servers:
  - url: https://quest-manas-pradhans-projects.vercel.app
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/users/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      tags: [Auth]
      summary: Update current user profile
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ===========================================
  # Workspaces
  # ===========================================
...
  /api/workspaces/{workspaceId}/projects:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'

    get:
      tags: [Projects]
      summary: List projects in workspace
      operationId: listProjects
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Projects]
      summary: Create a new project
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Validation error or key taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'

    get:
      tags: [Projects]
      summary: Get project details
      operationId: getProject
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Projects]
      summary: Update project
      operationId: updateProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Projects]
      summary: Delete project
      operationId: deleteProject
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only admins can delete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===========================================
  # Issues
  # ===========================================
  /api/projects/{projectId}/issues:
    parameters:
      - $ref: '#/components/parameters/ProjectId'

    get:
      tags: [Issues]
      summary: List issues in project
      operationId: listIssues
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/IssueStatus'
        - name: assigneeId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/IssueType'
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/IssuePriority'
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Issue'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Issues]
      summary: Create a new issue
      operationId: createIssue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequest'
      responses:
        '201':
          description: Issue created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: List user notifications
      operationId: listNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/workspaces/{workspaceId}/invitations:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'

    get:
      tags: [Invitations]
      summary: List pending invitations
      operationId: listInvitations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invitation'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Invitations]
      summary: Send an invitation
      operationId: createInvitation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
      responses:
        '201':
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationResponse'
        '400':
          description: User already member or invitation pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only owners and admins can invite
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: JWT or session token
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
      description: Workspace ID (CUID)
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
      description: Project ID (CUID)
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
    Forbidden:
      description: Not authorized to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "You don't have permission to perform this action"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Not found"
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
          format: email
        image:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    UserSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
        image:
          type: string
          nullable: true

    # ===========================================
    # Workspace Schemas
    # ===========================================
...
    WorkspaceRole:
      type: string
      enum: [ADMIN, DEVELOPER, TESTER, GUEST]
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        key:
          type: string
          description: Project key for issue IDs (e.g., "QUEST")
        description:
          type: string
          nullable: true
        workspaceId:
          type: string
        leadId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required: [name, key]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
        key:
          type: string
          minLength: 2
          maxLength: 10
          pattern: '^[A-Z0-9]+$'
          description: Uppercase letters and numbers only
        description:
          type: string
          maxLength: 500
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
        key:
          type: string
          minLength: 2
          maxLength: 10
        description:
          type: string
          maxLength: 500
        leadId:
          type: string
          nullable: true
    ProjectResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Project'

    # ===========================================
    # Issue Schemas
    # ===========================================
    IssueType:
      type: string
      enum: [EPIC, STORY, TASK, BUG]
    IssueStatus:
      type: string
      enum: [BACKLOG, TODO, IN_PROGRESS, IN_REVIEW, DONE, CANCELLED]
    IssuePriority:
      type: string
      enum: [URGENT, HIGH, MEDIUM, LOW, NONE]
    Issue:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
          description: Human-readable key (e.g., "QUEST-123")
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          $ref: '#/components/schemas/IssueType'
        status:
          $ref: '#/components/schemas/IssueStatus'
        priority:
          $ref: '#/components/schemas/IssuePriority'
        projectId:
          type: string
        assigneeId:
          type: string
          nullable: true
        reporterId:
          type: string
        dueDate:
          type: string
          format: date-time
          nullable: true
        order:
          type: integer
        parentId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    IssueWithRelations:
      allOf:
        - $ref: '#/components/schemas/Issue'
        - type: object
          properties:
            assignee:
              $ref: '#/components/schemas/UserSummary'
              nullable: true
            reporter:
              $ref: '#/components/schemas/UserSummary'
            labels:
              type: array
              items:
                type: object
                properties:
                  label:
                    $ref: '#/components/schemas/Label'
            subtasks:
              type: array
              items:
                $ref: '#/components/schemas/Subtask'
    Subtask:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        title:
          type: string
        type:
          $ref: '#/components/schemas/IssueType'
        status:
          $ref: '#/components/schemas/IssueStatus'
        priority:
          $ref: '#/components/schemas/IssuePriority'
    Label:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
          description: Hex color code
    CreateIssueRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 10000
        type:
          $ref: '#/components/schemas/IssueType'
          default: TASK
        priority:
          $ref: '#/components/schemas/IssuePriority'
          default: MEDIUM
        assigneeId:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        parentId:
          type: string
          nullable: true
          description: For creating subtasks
        labelIds:
          type: array
          items:
            type: string
...
    IssueResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/IssueWithRelations'
...
    NotificationType:
      type: string
      enum: [ISSUE_ASSIGNED, ISSUE_STATUS_CHANGED, COMMENT_ADDED]
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/NotificationType'
        title:
          type: string
        message:
          type: string
          nullable: true
        link:
          type: string
          nullable: true
        userId:
          type: string
        actorId:
          type: string
          nullable: true
        issueId:
          type: string
          nullable: true
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
        actor:
          $ref: '#/components/schemas/UserSummary'
          nullable: true
        issue:
          type: object
          nullable: true
          properties:
            id:
              type: string
            key:
              type: string
            title:
              type: string

    # ===========================================
    # Invitation Schemas
    # ===========================================
    InvitationStatus:
      type: string
      enum: [PENDING, ACCEPTED, REJECTED, EXPIRED]
    Invitation:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/WorkspaceRole'
        workspaceId:
          type: string
        invitedById:
          type: string
        token:
          type: string
        status:
          $ref: '#/components/schemas/InvitationStatus'
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    CreateInvitationRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/WorkspaceRole'
          default: DEVELOPER
    InvitationResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Invitation'

    # ===========================================
    # Filter Schemas
    # ===========================================
...